Rename expected_conditions/conditions.py to .../general.py, update relevant import statements.
