+++ tests/core/player_tools/test_equipment.py
@@ -6,9 +6,9 @@ from selenium.webdriver.support.ui import WebDriverWait
 
 from components.core.character import armor, weapon
 from components.core.character.tabs import Tabs
-from expected_conditions.conditions import modal_finished_closing
-from expected_conditions.conditions import sorting_arrow_up, sorting_arrow_down
-from expected_conditions.conditions import table_cell_updated
+from expected_conditions.general import modal_finished_closing
+from expected_conditions.general import sorting_arrow_up, sorting_arrow_down
+from expected_conditions.general import table_cell_updated
 from factories.core.character.armor import ArmorFactory
 from utils import general as ut
